Refine ecommerce dashboard charts
